+++ festival/src/test/java/likelion/festival/domain/menu/PriceTest.java
@@ -11,7 +11,7 @@ class PriceTest {
     @DisplayName("메뉴 가격이 0보다 작거나 50,000보다 큰 경우에는 예외를 발생시킨다.")
     @ValueSource(longs = {-1, 50001})
     @ParameterizedTest
-    void fail_create_price_under_0_or_over_50000(final long price) {
+    void fail_create_price_under_0_or_over_50000(final int price) {
         // given
         // when
         // then
